automation - support for stable, edge, prerelease
and latest-any via ImagePolicy

Other image policies could be added if more environments are needed. All
of our automation will write to the `main` branch, except the dev policy
which is meant to be used on forks. (Pretend this repo is hosted in an
org, and developers all have their own forks! The dev environments maybe
should generally be derived to one of those forks.)

Instead of alpha/beta/test/dev we have opted for an "edge" channel which
is similar to prerelease, but can have those breaking changes or partial
features implemented only as a pre-release even after the 1.0.0 stable.