dev
The dev environment is for users in forks. It provides a way for dev
users to patch in their own git repository and is provided for example,
and for a place to target the `latest-any` image policy.

(It is not meant to be used in the base repo, as `latest-any` can be
very noisy and wouldn't make any sense to automate in the main branch.)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>