stable channel vs latest
"latest" picks the latest release build, and might or might not be fresh
at any given time based on where we are in the automation cycles.

"stable" can take cues from the OCI config about when a release is ready
for prime-time, or something, right now it is just based on having a
stable semver number which we do not have, so this will not function.

The intention is that none of these fields need to be manually upgraded,
someone just comes along pushes a tag at the right time, so releases can
go out. Maybe later, automation replaces those manual inputs too based
on signals from other automation deploy success or failure in earlier
stages of the automation.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>